+++ main.go
@@ -1,8 +1,12 @@
 package main
 
 import (
+	"fmt"
 	"os"
+
 	"github.com/cyberark/conjur-api-go/conjurapi"
+	"github.com/karrick/golf"
+	log "github.com/sirupsen/logrus"
 )
 
 func RetrieveSecret(variableName string) {
@@ -11,12 +15,14 @@ func RetrieveSecret(variableName string) {
 	conjur, err := conjurapi.NewClientFromEnvironment(config)
 
 	if err != nil {
-		printAndExit(err)
+		log.Errorf("Failed creating a Conjur client: %s\n", err.Error())
+		os.Exit(1)
 	}
 
 	value, err := conjur.RetrieveSecret(variableName)
 	if err != nil {
-		printAndExit(err)
+		log.Errorln(err.Error())
+		os.Exit(1)
 	}
 
 	os.Stdout.Write([]byte(value))
@@ -24,21 +30,30 @@ func RetrieveSecret(variableName string) {
 }
 
 func main() {
-	if len(os.Args) != 2 {
-		os.Stderr.Write([]byte("A variable name or version flag must be given as the first and only argument!"))
-		os.Exit(-1)
+	var help = golf.BoolP('h', "help", false, "show help")
+	var version = golf.BoolP('V', "version", false, "show version")
+	var verbose = golf.BoolP('v', "verbose", false, "be verbose")
+
+	golf.Parse()
+	args := golf.Args()
+
+	if *version {
+		fmt.Println(VERSION)
+		os.Exit(0)
+	}
+	if *help {
+		golf.Usage()
+		os.Exit(0)
+	}
+	if len(args) == 0 {
+		golf.Usage()
+		os.Exit(1)
 	}
 
-	singleArgument := os.Args[1]
-	switch singleArgument {
-	case "-v","--version":
-		os.Stdout.Write([]byte(VERSION))
-	default:
-		RetrieveSecret(singleArgument)
+	log.SetFormatter(&log.TextFormatter{DisableTimestamp: true, DisableLevelTruncation: true})
+	if *verbose {
+		log.SetLevel(log.DebugLevel)
 	}
-}
 
-func printAndExit(err error) {
-	os.Stderr.Write([]byte(err.Error()))
-	os.Exit(1)
+	RetrieveSecret(args[0])
 }